fix: removed console logs and fixed group policy invoke in web
